refactor: renaming FullScreenViewer & ExplorerWindow for clarity; rearranging component folders
