+++ app/components/artifacts/Terminal.jsx
