+++ app/components/artifacts/Therac25Accidents.jsx
