+++ app/components/artifacts/Win95RecycleBin.jsx
