+++ app/components/artifacts/WindowsNTBSOD.jsx
