+++ app/components/artifacts/YTViewCounter.jsx
