+++ app/components/gallery/IncidentDescriptionCard.jsx
